Added more styling

+++ src/pages/CatShow.js
@@ -5,8 +5,13 @@ import { Button } from 'reactstrap'
 class CatShow extends Component {
 
   handleDelete = () => {
-    this.props.deleteCat(this.props.cat.id)
-  }
+      if(window.confirm("Are you sure you want to delete this profile?") === true) {
+        this.props.deleteCat(this.props.cat.id)
+        this.setState({ submitted: "true" })
+      } else {
+        this.setState({ submitted: "false" })
+      }
+    }
 
 
     render() {
@@ -17,9 +22,13 @@ class CatShow extends Component {
                 <p>Age: {this.props.cat && this.props.cat.age}</p>
                 <p>Enjoys: {this.props.cat && this.props.cat.enjoys}</p>
                 <img src={this.props.cat && this.props.cat.image} />
+                <div className="Buttons">
                 <NavLink to={`/catedit/${this.props.cat && this.props.cat.id}`}> <Button> Edit this Cat's Profile! </Button></NavLink>
                 <NavLink to ={"/catindex"}> <Button onClick={this.handleDelete}> Delete this Cat's Profile </Button> </NavLink>
+                <NavLink to={`/`}><Button>Go Back to Homepage</Button></NavLink>
+                </div>
               </div>
+
             </>
         )
     }
